Standardize InputBlocks components
